information_get/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pytest>=7.0.0
python-dateutil>=2.8.2
httpx>=0.24.0
